+++ tests/integration/test_cluster.py
@@ -1,7 +1,7 @@
 import json
 import os
 
-from .common import setup_cluster, exec_cmd, default_cluster, default_cluster_with_plugins  # noqa: F401
+from .common import setup_cluster, exec_cmd, default_cluster  # noqa: F401
 
 
 def test_cluster_list(default_cluster):
@@ -40,7 +40,7 @@ def test_empty_cluster_list():
     assert out == '[]\n'
 
 
-def test_cluster_setup_non_superuser(default_cluster_with_plugins):
+def test_cluster_setup_non_superuser(default_cluster):
     username = 'nonsuperuser'
     password = 'nonsuperpassword'
 
@@ -49,11 +49,10 @@ def test_cluster_setup_non_superuser(default_cluster_with_plugins):
     exec_cmd(['dcos', 'security', 'org', 'users',
               'create', username, '--password', password])
 
-    code, out, err = exec_cmd(['dcos', 'cluster', 'setup', default_cluster_with_plugins['dcos_url'],
+    code, out, err = exec_cmd(['dcos', 'cluster', 'setup', default_cluster['dcos_url'],
                                '--username', username, '--password', password])
     assert code == 0
     assert out == ""
-    assert err == ""
 
 
 def test_cluster_setup_with_acs_token_env(default_cluster):
@@ -63,7 +62,6 @@ def test_cluster_setup_with_acs_token_env(default_cluster):
     code, out, err = exec_cmd(['dcos', 'cluster', 'setup', default_cluster['dcos_url']], env=env)
     assert code == 0
     assert out == ""
-    assert err == ""
 
     code, out, _ = exec_cmd(['dcos', 'config', 'show', 'core.dcos_acs_token'])
     assert code == 0