+++ tests/integration/test_corecli.py
@@ -1,18 +1,23 @@
 import json
 import os
 
-from .common import exec_cmd, default_cluster, default_cluster_with_plugins  # noqa: F401
+from .common import exec_cmd, default_cluster  # noqa: F401
 
 import pytest
 
 
 @pytest.mark.skipif(os.environ.get('DCOS_TEST_CORECLI') is None, reason="no core CLI bundle")
 def test_extract_core(default_cluster):
+    code, out, err = exec_cmd(['dcos', 'plugin', 'remove', 'dcos-core-cli'])
+    assert code == 0
+    assert out == ''
+    assert err == ''
+
     cmd = ['dcos', 'package', 'describe', 'dcos-core-cli']
 
     code, out, err = exec_cmd(cmd)
     assert code == 0
-    assert err == "Extracting \"dcos-core-cli\"...\n"
+    assert err.startswith("Extracting \"dcos-core-cli\"...\n")
 
     pkgInfo = json.loads(out)
     assert pkgInfo['package']['name'] == 'dcos-core-cli'
@@ -25,8 +30,7 @@ def test_extract_core(default_cluster):
     assert pkgInfo['package']['name'] == 'dcos-core-cli'
 
 
-
-def test_update_core(default_cluster_with_plugins):
+def test_update_core(default_cluster):
     cmds = [
         ['dcos', 'package', 'install', 'dcos-core-cli'],
         ['dcos', 'package', 'install', 'dcos-core-cli', '--cli'],