+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
   * Introduce a strict deprecation mode (DCOS_CLI_STRICT_DEPRECATIONS=1)
   * Displayed variant when running dcos --version
+  * Always auto-install core and EE plugins on `cluster setup`